Update properties.py
hours support now operational

fix was that i forgot to actually update/toggle my boolean toggleable
got too caught up in using the ternary

need to figure out how to handle hours with zero activity. they're obv not important, so add clause that if not exist, skip to none available? valid, but doesn't cost that much to have them real and gathering dust, since nothing to index.
currently they are not added to the main hours dict, but subject to change.