Update properties.py
restaurant properties input mostly done. minor change so that convert_properties no longer only operates on the main restaurants list, and can be applied to other lists / splices as necessary

prompted input is mostly online, but
a. eval has to be able to access variables, so str need to be quoted to be recognised - the program breaks otherwise
b. it's difficult to manage overwrite access/protection using a nested list where $name, the differentiating value, is nested. if i generate a list of those, i need to update them, at which point i might as well use a dictionary or something. so i gotta update the rest of the code to work with that.

need to clean up code/properly document a bit